+++ package.json
@@ -2,10 +2,10 @@
   "name": "task-manager",
   "version": "1.0.0",
   "description": "Assignment 1 for backend engineering launchpad",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js",
+    "start": "node app.js",
+    "dev": "nodemon app.js",
     "test": "jest"
   },
   "author": "Airtribe",
